Refactor methods in GitLsTree
